1.对于待审批的记录，添加一个【详情】快捷操作ICON 2.点取消处理未上传附件的情况

+++ src/routes/Toponymy/Forms/BridgeForm.js
@@ -568,7 +568,11 @@ class BridgeForm extends Component {
         onCancel() {},
       });
     } else {
-      this.deleteUploadFiles(this.removeFileInfo);
+      if (this.removeFileInfo) {
+        this.deleteUploadFiles(this.removeFileInfo);
+      }else{
+        this.backToSearch();
+      }
     }
   }
