+++ src/routes/Toponymy/Forms/SettlementForm.js
@@ -563,7 +563,11 @@ class SettlementForm extends Component {
         onCancel() {},
       });
     } else {
-      this.deleteUploadFiles(this.removeFileInfo);
+      if (this.removeFileInfo) {
+        this.deleteUploadFiles(this.removeFileInfo);
+      }else{
+        this.backToSearch();
+      }
     }
   }
